move eventstore tests to new package
