+++ cfxdb/tests/eventstore/__init__.py
@@ -1,0 +1,6 @@
+##############################################################################
+#
+#                        Crossbar.io FX
+#     Copyright (C) Crossbar.io Technologies GmbH. All rights reserved.
+#
+##############################################################################
